moved settings to project root so it can be shared between the two demos.
