+++ demo-angular/src/app/pages/home/home-page.component.ts
@@ -32,7 +32,7 @@ import {
 // the Mapbox access token and the Mapbox map style to use are set
 // in the config.ts file.
 
-import { SETTINGS } from '../../config';
+import { SETTINGS } from '../../../../../mapbox_config';
 
 // ------------------------------------------------------------------------------
